+++ noops/package/helm.py
@@ -30,7 +30,7 @@ import shutil
 import re
 from typing import List
 import yaml
-from . import helper
+from .. import helper
 
 class Helm():
     """
